+++ lib/src/features/vn_item/presentation/vn_item_grid_cover.dart
@@ -1,10 +1,14 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:visibility_detector/visibility_detector.dart';
 import 'package:vndb_lite/src/app.dart';
 import 'package:vndb_lite/src/common_widgets/generic_image_error.dart';
 import 'package:vndb_lite/src/features/settings/presentation/settings_general_state.dart';
 import 'package:vndb_lite/src/features/vn/domain/others.dart';
+import 'package:vndb_lite/src/features/vn_item/presentation/vn_item_grid_cover_censor_notifier.dart';
 import 'package:vndb_lite/src/util/alt_provider_reader.dart';
 import 'package:vndb_lite/src/util/custom_cache_manager.dart';
 import 'package:vndb_lite/src/util/responsive.dart';
@@ -15,12 +19,14 @@ class VnItemGridCover extends StatefulWidget {
     required this.vnId,
     required this.image,
     required this.isGridView,
+    required this.labelCode,
     this.title,
   });
 
   final String vnId;
   final VnImage? image;
   final bool isGridView;
+  final String labelCode;
 
   /// Debugging purpose only.
   final String? title;
@@ -44,32 +50,34 @@ class VnItemGridCover extends StatefulWidget {
 class _VnItemGridCoverState extends State<VnItemGridCover> {
   late final String? _coverUrl;
   late final bool _hasCover;
-  late final bool _isSystematicallyCensored;
+  late final String _blurId;
+
+  bool _isSystematicallyCensored = false;
 
   @override
   void initState() {
     super.initState();
     _coverUrl = widget.image?.thumbnail;
     _hasCover = widget.image != null && _coverUrl != null;
+    _blurId = widget.vnId + widget.labelCode + App.currentRoute;
 
     if (_hasCover) {
       final settings = ref_.read(settingsGeneralStateProvider);
       if (settings.showCoverCensor && _vnMatchCensorRequirement) {
         _isSystematicallyCensored = true;
-        return;
       }
     }
 
-    _isSystematicallyCensored = false;
+    SchedulerBinding.instance.addPostFrameCallback((_) {
+      ref_
+          .read(vnItemGridCoverCensorNotifierProvider(_blurId).notifier)
+          .set(_isSystematicallyCensored);
+    });
   }
 
   @override
-  void dispose() async {
-    debugPrint('Disposing ${widget.title}');
-    VnItemGridCover.coverBlurToggle.remove(widget.vnId);
-
-    // if (App.isInSearchScreen) CachedNetworkImage.evictFromCache(_coverUrl ?? "");
-    // PaintingBinding.instance.imageCache.clear();
+  void dispose() {
+    DefaultCacheManager().emptyCache();
     super.dispose();
   }
 
@@ -77,41 +85,45 @@ class _VnItemGridCoverState extends State<VnItemGridCover> {
     return (widget.image?.sexual ?? 0) >= 1 || (widget.image?.violence ?? 0) >= 1;
   }
 
-  void _toggleBlur() {
-    final value = VnItemGridCover.coverBlur[widget.vnId];
-    setState(() => VnItemGridCover.coverBlur[widget.vnId] = !(value ?? _isSystematicallyCensored));
-  }
-
   @override
   Widget build(BuildContext context) {
-    final isCensor = VnItemGridCover.coverBlur[widget.vnId] ?? _isSystematicallyCensored;
-    VnItemGridCover.coverBlurToggle[widget.vnId] = _toggleBlur;
-    final cover = ConstrainedBox(
-      constraints: const BoxConstraints(
-        minHeight: VnItemGridCover.minHeightSize,
-        minWidth: VnItemGridCover.minWidthSize,
-      ),
-      child: CachedNetworkImage(
-        imageUrl: (_hasCover) ? (_coverUrl ?? '') : '',
-        width: (widget.isGridView) ? double.infinity : null,
-        height: (widget.isGridView) ? null : VnItemGridCover.nonGridViewHeight,
-        placeholder:
-            (_, _) => SizedBox.square(
-              dimension: VnItemGridCover.sizes[widget.vnId] ?? VnItemGridCover.placeHolderSize,
-            ),
-        fit: BoxFit.cover,
-        errorWidget: (_, url, error) => const GenericErrorImage(),
-        errorListener: null,
-        cacheManager: (!App.isInSearchScreen) ? CustomCacheManager() : null,
-        cacheKey:
-            (!App.isInSearchScreen)
-                ? (isCensor)
-                    ? "CENSORED-PREVIEW-${widget.vnId}"
-                    : "PREVIEW-${widget.vnId}"
-                : null,
-        maxHeightDiskCache: (isCensor) ? 15 : null,
-        maxWidthDiskCache: (isCensor) ? 15 : null,
-      ),
+    final cover = Consumer(
+      builder: (context, ref, child) {
+        final isCensor =
+            ref.watch(vnItemGridCoverCensorNotifierProvider(_blurId)) ?? _isSystematicallyCensored;
+
+        // return FutureBuilder(
+        //   future: CustomCacheManager().getSingleFile(widget.image!.thumbnail!),
+        //   builder: (BuildContext context, AsyncSnapshot snapshot) {
+        //     if (snapshot.connectionState == ConnectionState.waiting) {
+        //       return CircularProgressIndicator();
+        //     }
+        //     return Image.file(snapshot.data, fit: BoxFit.cover);
+        //   },
+        // );
+
+        return CachedNetworkImage(
+          imageUrl: (_hasCover) ? (_coverUrl ?? '') : '',
+          width: (widget.isGridView) ? double.infinity : null,
+          height: (widget.isGridView) ? null : VnItemGridCover.nonGridViewHeight,
+          placeholder:
+              (_, _) => SizedBox.square(
+                dimension: VnItemGridCover.sizes[widget.vnId] ?? VnItemGridCover.placeHolderSize,
+              ),
+          fit: BoxFit.cover,
+          errorWidget: (_, url, error) => const GenericErrorImage(),
+          errorListener: null,
+          cacheManager: (!App.isInSearchScreen) ? CustomCacheManager() : null,
+          cacheKey:
+              (!App.isInSearchScreen)
+                  ? (isCensor)
+                      ? "CENSORED-PREVIEW-${widget.vnId}"
+                      : "PREVIEW-${widget.vnId}"
+                  : null,
+          maxHeightDiskCache: (isCensor) ? 15 : null,
+          maxWidthDiskCache: (isCensor) ? 15 : null,
+        );
+      },
     );
 
     // * Exists only to fetch the size for the placeholder image to prevent